Fix canonical_projections.player_id overflow: INTEGER  BIGINT
Yahoo fallback IDs (-(yahoo_id)) for MLBAM-null players can exceed INT4
range (e.g., yahoo_id=91689018531). This caused NumericValueOutOfRange
errors for 9 players (Johnny Brito, JP France, Kenny Hernandez, etc.)
during canonical_projection_refresh.

Changes:
- backend/models.py: CanonicalProjection.player_id Column(Integer)  BigInteger
- scripts/migration_canonical_player_id_bigint.py: ALTER TABLE ... TYPE BIGINT

Run migration before next canonical_projection_refresh:
  railway run python scripts/migration_canonical_player_id_bigint.py

+++ backend/models.py
@@ -2043,7 +2043,8 @@ class CanonicalProjection(Base):
     projection_id = Column(String(36), nullable=False, unique=True, index=True)  # UUID
 
     # Identity — player_id links to player_id_mapping.mlbam_id (canonical int)
-    player_id = Column(Integer, nullable=False, index=True)
+    # BigInteger required: Yahoo fallback IDs like -(yahoo_id) can exceed INT4 range (~2.1B)
+    player_id = Column(BigInteger, nullable=False, index=True)
     player_type = Column(String(10), nullable=False)          # BATTER | PITCHER
     source_engine = Column(String(25), nullable=False)        # BAYESIAN | SAVANT_ADJUSTED | STATIC_BOARD | FALLBACK_MARCEL
     projection_date = Column(Date, nullable=False, index=True)
